organised test files in folders
